Move todo open counts into todo projection
